fix: api 변경에 따른 수정
