+++ lib/api.js
@@ -24,8 +24,14 @@ export const changeUserPassword = (body, header) =>
 export const searchUserByNickname = (nickname, header) =>
   axios.get(`${Config.baseUrl}/api/users/nickname/${nickname}`, header);
 
-// Team 관련
-export const getReadyTeam = (pageNumber,header) =>
+// 팀 관련
+export const getTeamInfo = (id, header) =>
+  axios.get(`${Config.baseUrl}/api/teams/${id}`, header);
+
+export const getTeamMemberInfo = (id, header) =>
+  axios.get(`${Config.baseUrl}/api/teams/${id}/members`, header);
+
+export const getReadyTeam = (pageNumber, header) =>
   axios.get(`${Config.baseUrl}/api/teams?page=${pageNumber}`, header);
 
 export const changeTeamStatus = (id, body, header) =>
@@ -34,11 +40,11 @@ export const changeTeamStatus = (id, body, header) =>
 export const createTeam = (body, header) =>
   axios.post(`${Config.baseUrl}/api/teams`, body, header);
 
-export const exitTeam = (id, header) =>
-  axios.delete(`${Config.baseUrl}/api/teams/${id}/members`, header);
+export const exitTeam = (body, header) =>
+  axios.post(`${Config.baseUrl}/api/teams/leave`, body, header);
 
-export const deleteTeam = (id, header) =>
-  axios.delete(`${Config.baseUrl}/api/teams/${id}`, header);
+export const deleteTeam = (header) =>
+  axios.delete(`${Config.baseUrl}/api/teams`, header);
 
 // 팀 초대 관련
 export const inviteTeam = (teamId, userId, header) =>
@@ -62,7 +68,7 @@ export const refusalInvitation = (invitationId, header) =>
     `${Config.baseUrl}/api/invitations/${invitationId}/refuse`,
     header,
   );
-  
+
 // auth 관련
 export const register = ({
   email,
