+++ lib/createRequestSaga.js
@@ -17,7 +17,7 @@ export default function createRequestSaga(type, request) {
       const response = yield call(request, action.payload);
       yield put({
         type: SUCCESS,
-        payload: response.headers,
+        payload: response.data,
       });
     } catch (e) {
       yield put({
